abstract stripe stuff to a concern
